anv/event: Use a 64-bit value
The immediate write from PIPE_CONTROL is 64-bits at least on BDW.  This
used to work on 64-bit archs because the compiler would align the following
anv_state struct up for us.  However, in 32-bit builds, they overlap and it
causes problems.

+++ src/vulkan/anv_private.h
@@ -1306,7 +1306,7 @@ struct anv_fence {
 };
 
 struct anv_event {
-   uint32_t                                     semaphore;
+   uint64_t                                     semaphore;
    struct anv_state                             state;
 };
